make fixtures importable downstream
