+++ conda/testing/fixtures.py
@@ -1,0 +1,32 @@
+import warnings
+import py
+import pytest
+
+from conda.common.compat import PY3
+from conda.gateways.disk.create import TemporaryDirectory
+from conda.core.subdir_data import SubdirData
+
+
+@pytest.fixture(autouse=True)
+def suppress_resource_warning():
+    '''
+    Suppress `Unclosed Socket Warning`
+
+    It seems urllib3 keeps a socket open to avoid costly recreation costs.
+
+    xref: https://github.com/kennethreitz/requests/issues/1882
+    '''
+    if PY3:
+        warnings.filterwarnings("ignore", category=ResourceWarning)
+
+
+@pytest.fixture(scope='function')
+def tmpdir(tmpdir, request):
+    tmpdir = TemporaryDirectory(dir=str(tmpdir))
+    request.addfinalizer(tmpdir.cleanup)
+    return py.path.local(tmpdir.name)
+
+
+@pytest.fixture(autouse=True)
+def clear_subdir_cache():
+    SubdirData.clear_cached_local_channel_data()
