+++ tests/conftest.py
@@ -2,21 +2,16 @@
 # Copyright (C) 2012 Anaconda, Inc
 # SPDX-License-Identifier: BSD-3-Clause
 
-from functools import partial
-import os
 import sys
-import warnings
 
-import py
-import pytest
-
-from conda.common.compat import PY3
-from conda.gateways.disk.create import TemporaryDirectory
-from conda.core.subdir_data import SubdirData
+from conda.testing.fixtures import (
+    suppress_resource_warning,
+    tmpdir,
+    clear_subdir_cache,
+)
 
 win_default_shells = ["cmd.exe", "powershell", "git_bash", "cygwin"]
 shells = ["bash", "zsh"]
-
 if sys.platform == "win32":
     shells = win_default_shells
 
@@ -29,27 +24,3 @@ def pytest_addoption(parser):
 def pytest_generate_tests(metafunc):
     if 'shell' in metafunc.fixturenames:
         metafunc.parametrize("shell", metafunc.config.option.shell)
-
-
-@pytest.fixture(autouse=True)
-def suppress_resource_warning():
-    '''
-Suppress `Unclosed Socket Warning`
-
-It seems urllib3 keeps a socket open to avoid costly recreation costs.
-
-xref: https://github.com/kennethreitz/requests/issues/1882
-'''
-    if PY3:
-        warnings.filterwarnings("ignore", category=ResourceWarning)
-
-@pytest.fixture(scope='function')
-def tmpdir(tmpdir, request):
-    tmpdir = TemporaryDirectory(dir=str(tmpdir))
-    request.addfinalizer(tmpdir.cleanup)
-    return py.path.local(tmpdir.name)
-
-
-@pytest.fixture(autouse=True)
-def clear_subdir_cache():
-    SubdirData.clear_cached_local_channel_data()
